restrcutured min-side
